// routes/logic/classes/interfaces/CustomerAddressInterface.ts
export interface CustomerAddressInterface {
    id:number;
    customerAddressType:string;
    customersNumber: string;
    customerAddressIsoCode: string;
    customerAddressCountryName: string;
    customerAddressStreet: string;
    customerAddressCity: string;
    customerAddressPostcode: string;
    customerAddressComment: string;
    customerAddressTaxCode: string;
    customerAddressNameAddr: string;
    customerAddressEmail: string;
    customerAddressPhone: string;
}

export interface CustomerAddressDataInterface {
    ID:number;
    ADDRESS_TYPE:string;
    CUSTOMERS_NUMBER:string;
    ADDRESS_ISO_CODE:string;
    ADDRESS_CRYNAME:string;
    ADDRESS_STREET:string;
    ADDRESS_CITY:string;
    ADDRESS_POSTCODE:string;
    ADDRESS_COMMENT:string;
    TAXCODE: string;
    NAME_ADDR: string;
    EMAIL: string;
    PHONE: string;
}